+++ src/fut/stream_then.rs
@@ -1,7 +1,7 @@
 use futures::{Async, Poll};
 
-use actor::Actor;
-use fut::{ActorFuture, ActorStream, IntoActorFuture};
+use crate::actor::Actor;
+use crate::fut::{ActorFuture, ActorStream, IntoActorFuture};
 
 /// A stream combinator which chains a computation onto each item produced by a
 /// stream.
@@ -50,7 +50,9 @@ where
     type Actor = S::Actor;
 
     fn poll(
-        &mut self, act: &mut S::Actor, ctx: &mut <S::Actor as Actor>::Context,
+        &mut self,
+        act: &mut S::Actor,
+        ctx: &mut <S::Actor as Actor>::Context,
     ) -> Poll<Option<U::Item>, U::Error> {
         if self.future.is_none() {
             let item = match self.stream.poll(act, ctx) {