+++ tests/derive.rs
@@ -1,10 +1,9 @@
-extern crate futures;
 #[macro_use]
 extern crate actix;
-extern crate tokio;
 
 use actix::{Actor, Context, Handler, System};
 use futures::{future, Future};
+use std::ops::Mul;
 
 #[derive(Message)]
 struct Empty;
@@ -28,7 +27,7 @@ fn response_derive_empty() {
         let addr = EmptyActor.start();
         let res = addr.send(Empty);
 
-        tokio::spawn(res.then(|res| {
+        actix::spawn(res.then(|res| {
             match res {
                 Ok(result) => assert!(result == ()),
                 _ => panic!("Something went wrong"),
@@ -37,7 +36,8 @@ fn response_derive_empty() {
             System::current().stop();
             future::result(Ok(()))
         }));
-    });
+    })
+    .unwrap();
 }
 
 #[derive(Message)]
@@ -54,7 +54,9 @@ impl Handler<SumResult> for SumResultActor {
     type Result = Result<usize, ()>;
 
     fn handle(
-        &mut self, message: SumResult, _context: &mut Context<Self>,
+        &mut self,
+        message: SumResult,
+        _context: &mut Context<Self>,
     ) -> Self::Result {
         Ok(message.0 + message.1)
     }
@@ -66,7 +68,7 @@ pub fn derive_result() {
         let addr = SumResultActor.start();
         let res = addr.send(SumResult(10, 5));
 
-        tokio::spawn(res.then(|res| {
+        actix::spawn(res.then(|res| {
             match res {
                 Ok(result) => assert!(result == Ok(10 + 5)),
                 _ => panic!("Something went wrong"),
@@ -75,7 +77,8 @@ pub fn derive_result() {
             System::current().stop();
             future::result(Ok(()))
         }));
-    });
+    })
+    .unwrap();
 }
 
 #[derive(Message)]
@@ -102,7 +105,7 @@ pub fn response_derive_one() {
         let addr = SumOneActor.start();
         let res = addr.send(SumOne(10, 5));
 
-        tokio::spawn(res.then(|res| {
+        actix::spawn(res.then(|res| {
             match res {
                 Ok(result) => assert!(result == 10 + 5),
                 _ => panic!("Something went wrong"),
@@ -111,5 +114,90 @@ pub fn response_derive_one() {
             System::current().stop();
             future::result(Ok(()))
         }));
-    });
+    })
+    .unwrap();
+}
+
+#[derive(MessageResponse, PartialEq)]
+struct MulRes(usize);
+
+#[derive(Message)]
+#[rtype(MulRes)]
+struct MulOne(usize, usize);
+
+struct MulOneActor;
+
+impl Actor for MulOneActor {
+    type Context = Context<Self>;
+}
+
+impl Handler<MulOne> for MulOneActor {
+    type Result = MulRes;
+
+    fn handle(&mut self, message: MulOne, _context: &mut Context<Self>) -> Self::Result {
+        MulRes(message.0 * message.1)
+    }
+}
+
+#[test]
+pub fn derive_response_one() {
+    System::run(|| {
+        let addr = MulOneActor.start();
+        let res = addr.send(MulOne(10, 5));
+
+        actix::spawn(res.then(|res| {
+            match res {
+                Ok(result) => assert!(result == MulRes(10 * 5)),
+                _ => panic!("Something went wrong"),
+            }
+
+            System::current().stop();
+            future::result(Ok(()))
+        }));
+    })
+    .unwrap();
+}
+
+#[derive(MessageResponse, PartialEq)]
+struct MulAny<T: 'static + Mul>(T);
+
+#[derive(Message)]
+#[rtype(result = "MulAny<usize>")]
+struct MulAnyOne(usize, usize);
+
+struct MulAnyOneActor;
+
+impl Actor for MulAnyOneActor {
+    type Context = Context<Self>;
+}
+
+impl Handler<MulAnyOne> for MulAnyOneActor {
+    type Result = MulAny<usize>;
+
+    fn handle(
+        &mut self,
+        message: MulAnyOne,
+        _context: &mut Context<Self>,
+    ) -> Self::Result {
+        MulAny(message.0 * message.1)
+    }
+}
+
+#[test]
+pub fn derive_response_two() {
+    System::run(|| {
+        let addr = MulAnyOneActor.start();
+        let res = addr.send(MulAnyOne(10, 5));
+
+        actix::spawn(res.then(|res| {
+            match res {
+                Ok(result) => assert!(result == MulAny(10 * 5)),
+                _ => panic!("Something went wrong"),
+            }
+
+            System::current().stop();
+            future::result(Ok(()))
+        }));
+    })
+    .unwrap();
 }